requirements.txt
plotext
python-dotenv
googlemaps